move datalog fact generation into a legacy module
